Add steal phone number steal message
